chapter-16-rag/requirements.txt
-e openai>=1.0
chromadb>=0.4
python-dotenv>=1.0
